Extract search algorithm into file
