+++ frontend/lib/js/concept-trees/globalTreeStoreHelper.js
@@ -7,6 +7,8 @@ import type { TreesT } from "./reducer";
 import { tablesWithDefaults } from "../model/table";
 import { selectsWithDefaults } from "../model/select";
 
+import { findConcepts } from "./search";
+
 // Globally store the huge (1-5 MB) trees for read only
 // - keeps the redux store free from huge data
 window.conceptTrees = {};
@@ -130,7 +132,7 @@ export const hasConceptChildren = (node: ConceptT): boolean => {
   This is async because ... we might want to parallelize this very soon,
   as there are up to 200k concepts that need to be searched.
 */
-export const search = async (query: string) => {
+export const globalSearch = async (query: string) => {
   const lowerQuery = query.toLowerCase();
 
   const result = Object.keys(window.conceptTrees).reduce(
@@ -142,82 +144,4 @@ export const search = async (query: string) => {
   );
 
   return result;
-};
-
-const doesQueryMatchNode = (node, query) => {
-  return (
-    node.label.toLowerCase().includes(query) ||
-    (node.description && node.description.toLowerCase().includes(query)) ||
-    (node.additionalInfos &&
-      node.additionalInfos
-        .map(({ value }) => value)
-        .join("")
-        .toLowerCase()
-        .includes(query))
-  );
-};
-
-/*
-  This is a recursive algorithm to search through the trees
-  It counts results and stores the count in "intermediateResult".
-
-  The code might look a little bit "clumsy", but
-  - it's been optimized to _not_ use object spread, because that slowed it down
-  - it's been optimized to have a time complexity of O(n)
-
-  treeId: the root tree id
-  nodeId: the id of the concept node, because node doesn't include it itself
-  node: the current node to check for match,
-    includes all information needed, eg: label, description, additionalInfos and children
-  query: the search query
-  intermediateResult: to avoid building new objects in every iteration, we carry
-    this object through the recursion and define new properties as we go (side effects)
-*/
-const findConcepts = (
-  treeId: string,
-  nodeId: ConceptIdT,
-  node: ConceptT,
-  query: string,
-  intermediateResult: { [ConceptIdT]: number } = {}
-) => {
-  // !node normall shouldn't happen.
-  // It happens when window.conceptTrees doesn't contain a node
-  // that was somewhere specified as a child.
-  // TODO: Stronger contract with API on what shape of data is allowed
-  if (!node) return intermediateResult;
-
-  const isNodeIncluded = doesQueryMatchNode(node, query);
-
-  // Early return if there are no children
-  if (!node.children) {
-    if (isNodeIncluded) {
-      intermediateResult[nodeId] = 1;
-    }
-
-    return intermediateResult;
-  }
-
-  // Count node as 1 already, if it matches
-  let sum = isNodeIncluded ? 1 : 0;
-
-  for (let child of node.children) {
-    const result = findConcepts(
-      treeId,
-      child,
-      window.conceptTrees[treeId][child],
-      query,
-      intermediateResult
-    );
-
-    sum += result[child] || 0;
-  }
-
-  if (sum !== 0) {
-    intermediateResult[nodeId] = sum;
-  }
-
-  // If sum === 0, node is left out from the result,
-  // because it doesn't match itself and no child matches
-
-  return intermediateResult;
 };
